Refactor OAuth integration by removing BaseOAuthDriver and updating drivers for improved structure
- Deleted the `BaseOAuthDriver` class, which provided default implementations for email restrictions and additional parameters, as it was no longer needed.
- Updated `OAuthGoogleDriver`, `OAuthGoogleOneTapDriver`, `OAuthStripeDriver`, and `OAuthTelegramDriver` to implement their respective interfaces directly, enhancing clarity and reducing unnecessary inheritance.
- Introduced a new trait `HasEmailRestrictions` to manage email restrictions, ensuring that only drivers that require this functionality implement it.
- Adjusted the `OAuthInviteService` to utilize capability-based checks for email restrictions, improving the overall design and maintainability of the OAuth flow.

These changes aim to streamline the OAuth driver architecture, promote better separation of concerns, and enhance the overall maintainability of the codebase.
