+++ api/app/Integrations/OAuth/Drivers/Contracts/OAuthDriver.php
@@ -16,21 +16,4 @@ interface OAuthDriver
      * Get the appropriate scopes for a given intent (auth or integration).
      */
     public function getScopesForIntent(string $intent): array;
-
-    /**
-     * Set email restrictions/hints for the OAuth flow
-     * Default implementation should be no-op for drivers that don't support it
-     */
-    public function setEmailRestrictions(array $emails): self;
-
-    /**
-     * Check if this driver supports email restrictions
-     */
-    public function supportsEmailRestrictions(): bool;
-
-    /**
-     * Set additional OAuth parameters (generic method for customization)
-     * This allows drivers to add custom parameters to the OAuth redirect
-     */
-    public function setAdditionalParameters(array $parameters): self;
 }
