+++ api/app/Integrations/OAuth/Drivers/Contracts/SupportsEmailRestrictions.php
@@ -1,0 +1,16 @@
+<?php
+
+namespace App\Integrations\OAuth\Drivers\Contracts;
+
+interface SupportsEmailRestrictions
+{
+    /**
+     * Set email restrictions/hints for the OAuth flow
+     */
+    public function setEmailRestrictions(array $emails): self;
+
+    /**
+     * Check if this driver supports email restrictions
+     */
+    public function supportsEmailRestrictions(): bool;
+}
