+++ api/app/Integrations/OAuth/Drivers/OAuthGoogleOneTapDriver.php
@@ -3,14 +3,12 @@
 namespace App\Integrations\OAuth\Drivers;
 
 use App\Integrations\OAuth\Drivers\Contracts\WidgetOAuthDriver;
-use App\Integrations\OAuth\Drivers\Traits\SupportsEmailRestrictions;
 use Illuminate\Support\Facades\Cache;
 use Illuminate\Support\Facades\Http;
 use Laravel\Socialite\Contracts\User;
 
-class OAuthGoogleOneTapDriver extends BaseOAuthDriver implements WidgetOAuthDriver
+class OAuthGoogleOneTapDriver implements WidgetOAuthDriver
 {
-    use SupportsEmailRestrictions;
 
     public function getRedirectUrl(): string
     {
